Add Color::None option to allow only changing foreground or background
This change allows to only specify a foreground OR bakground color. The
reason for the change is the use of the SelectView with styled text.
Normally the SelectView will highlight the selection. But styled text
requires the specification of a foreground and background color. Now
with styled text in a SelectView, the highlight is broken. To allow the
highlight but also support custom text color, the change is added. Now
you can specify a foreground color for the text and set the background
color to Color::None. The background color is then determined by the
upper logic, for example by the SelectView if the highlight is added.