Add `Effect::Dim` (#580)
The "dim" or "faint" effect is set by the ANSI escape code `ESC[2m`, and supported in some terminals. It's more or less the opposite of the "bold" effect.

I couldn't get `ncurses`/`pancurses` to render anything differently, even though there's apparently an `A_DIM` constant; I'm not sure why. It renders with `crossterm` and `termion`.