enable iterating SelectView items in reverse (#819)
* add iterator trait aliases for SelectView::iter/iter_mut/try_iter_mut

Instead of returning `impl Iterator<...>`, they can effectively return
`impl Iterator<...> + DoubleEndedIterator + ExactSizeIterator`, because
that is what the underlying iterators actually implement.

This lets you iterate in reverse if you want, which isn't possible if it
just returns `impl Iterator`.

Unfortunately the syntax for doing this is very verbose.

I believe this should be seamlessly backwards-compatible, since it's
returning a type with the strict superset of traits it previously did,
and the previous return type was an opaque type.

* return multiple traits instead of making our own supertrait